Subject [ Modify ] : 刪除更新重開機邏輯 Message : 調整刪除判斷邏輯

+++ app/src/main/java/com/ufistudio/hotelmediabox/pages/SplashActivity.kt
@@ -39,7 +39,7 @@ class SplashActivity : AppCompatActivity() {
 
     override fun onStart() {
         super.onStart()
-        if (FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME)?.exists()!!) {
+        if (FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME) != null && FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME)?.exists() == true) {
             FileUtils.getFileFromStorage(TAG_DEFAULT_APK_NAME)?.delete()
             MiscUtils.reboot(this)
         }
